Restructure ArchivePage functions to be cleaner
